refactor(mcp): extract session expiry classifier
